# openapi.yaml
openapi: 3.0.3
info:
  title: Qubic Explorer API
  description: API for exploring the Qubic network — transactions, addresses, epochs, smart contracts, analytics, and more.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/stats/chart/tx-volume:
    get:
      tags: [Stats]
      summary: Transaction volume chart data
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [day, week, month], default: day }
      responses:
        '200':
          description: Chart data points
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChartDataPointDto' }
  /api/stats/top-addresses:
    get:
      tags: [Stats]
      summary: Top addresses by volume
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: epoch
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Top addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TopAddressDto' }
  /api/stats/rich-list:
    get:
      tags: [Stats]
      summary: Rich list with pagination
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Rich list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RichListDto' }
...
  /api/stats/whale-alerts:
    get:
      tags: [Stats]
      summary: Large transfer alerts
      parameters:
        - name: threshold
          in: query
          schema: { type: integer, format: uint64, default: 10000000000 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Whale alerts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WhaleAlertDto' }
  /api/stats/smart-contract-usage:
    get:
      tags: [Stats]
      summary: Smart contract usage stats
      parameters:
        - name: epoch
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: SC usage
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SmartContractUsageDto' }
  /api/stats/active-addresses:
    get:
      tags: [Stats]
      summary: Active address trends
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [epoch, day], default: epoch }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Active address trends
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActiveAddressTrendDto' }
  /api/stats/new-vs-returning:
    get:
      tags: [Stats]
      summary: New vs returning addresses per epoch
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: New vs returning
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NewVsReturningDto' }
  /api/stats/exchange-flows:
    get:
      tags: [Stats]
      summary: Exchange inflow/outflow data
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Exchange flows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExchangeFlowDto' }
  /api/stats/exchange-senders:
    get:
      tags: [Stats]
      summary: Addresses sending to exchanges
      parameters:
        - name: epochs
          in: query
          schema: { type: integer, default: 5 }
        - name: minAmount
          in: query
          schema: { type: integer, format: uint64, default: 1000000000 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: depth
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: Exchange senders
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExchangeSendersDto' }
...
  /api/stats/holder-distribution/extended:
    get:
      tags: [Stats]
      summary: Holder distribution with history
      parameters:
        - name: historyLimit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Extended holder distribution
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HolderDistributionExtendedDto' }
  /api/stats/holder-distribution/history:
    get:
      tags: [Stats]
      summary: Historical holder distribution snapshots
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: History snapshots
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HolderDistributionHistoryDto' }
  /api/stats/avg-tx-size:
    get:
      tags: [Stats]
      summary: Average transaction size trends
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [epoch, day], default: epoch }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Avg TX size trends
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AvgTxSizeTrendDto' }
  /api/stats/network-stats/history:
    get:
      tags: [Stats]
      summary: Network stats history
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: History
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NetworkStatsHistoryDto' }
  /api/stats/network-stats/extended:
    get:
      tags: [Stats]
      summary: Network stats with history
      parameters:
        - name: historyLimit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Extended network stats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkStatsExtendedDto' }
...
  /api/stats/burn-stats/history:
    get:
      tags: [Stats]
      summary: Burn stats history
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Burn stats history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BurnStatsHistoryDto' }
  /api/stats/burn-stats/extended:
    get:
      tags: [Stats]
      summary: Burn stats with history
      parameters:
        - name: historyLimit
          in: query
          schema: { type: integer, default: 30 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Extended burn stats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BurnStatsExtendedDto' }
...
  /api/stats/computor-revenue/{epoch}:
    get:
      tags: [Stats]
      summary: Computor revenue for specific epoch
      parameters:
        - name: epoch
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Computor revenue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComputorRevenueDto' }
  /api/stats/computor-revenue/{epoch}/simulate:
    get:
      tags: [Stats]
      summary: Simulate computor revenue with custom tick cutoffs
      description: >
        Recalculates revenue on-the-fly using data up to the specified tick for each score category.
        Omitted tick parameters default to the latest network tick.
      parameters:
        - name: epoch
          in: path
          required: true
          schema: { type: integer }
        - name: txTick
          in: query
          schema: { type: integer, format: uint64 }
          description: Calculate TX scores using ticks up to this tick number
        - name: voteTick
          in: query
          schema: { type: integer, format: uint64 }
          description: Calculate vote scores using transactions up to this tick number
        - name: miningTick
          in: query
          schema: { type: integer, format: uint64 }
          description: Calculate mining scores using transactions up to this tick number
      responses:
        '200':
          description: Simulated revenue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComputorRevenueSimulationDto' }
        '404':
          description: Computor list unavailable for this epoch
  /api/stats/qearn/backfill:
    post:
      tags: [Stats, Admin]
      summary: Backfill qearn stats (requires admin API key)
      responses:
        '200':
          description: Success
      security:
        - AdminApiKey: []
  /api/stats/emission/backfill:
    post:
      tags: [Stats, Admin]
      summary: Backfill emission stats (requires admin API key)
      responses:
        '200':
          description: Success
      security:
        - AdminApiKey: []

  # ── Miner Flow ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  schemas:
...
    ChartDataPointDto:
      type: object
      properties:
        date: { type: string, format: date-time }
        txCount: { type: integer, format: uint64 }
        volume: { type: integer, format: uint64 }
    TopAddressDto:
      type: object
      properties:
        address: { type: string }
        label: { type: string, nullable: true }
        type: { type: string, nullable: true }
        sentVolume: { type: integer, format: uint64 }
        receivedVolume: { type: integer, format: uint64 }
        totalVolume: { type: integer, format: uint64 }
        sentCount: { type: integer }
        receivedCount: { type: integer }
        totalCount: { type: integer }
    SmartContractUsageDto:
      type: object
      properties:
        address: { type: string }
        name: { type: string }
        contractIndex: { type: integer, nullable: true }
        callCount: { type: integer, format: uint64 }
        totalAmount: { type: integer, format: uint64 }
        uniqueCallers: { type: integer }
    RichListDto:
      type: object
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/RichListEntryDto' }
        page: { type: integer }
        limit: { type: integer }
        totalCount: { type: integer, format: uint64 }
        totalPages: { type: integer }
        totalBalance: { type: number }
        snapshotEpoch: { type: integer }
...
    WhaleAlertDto:
      type: object
      properties:
        tickNumber: { type: integer, format: uint64 }
        epoch: { type: integer }
        txHash: { type: string }
        sourceAddress: { type: string }
        sourceLabel: { type: string, nullable: true }
        sourceType: { type: string, nullable: true }
        destAddress: { type: string }
        destLabel: { type: string, nullable: true }
        destType: { type: string, nullable: true }
        amount: { type: number }
        amountFormatted: { type: string }
        timestamp: { type: string, format: date-time }
    ActiveAddressTrendDto:
      type: object
      properties:
        epoch: { type: integer, nullable: true }
        date: { type: string, format: date-time, nullable: true }
        uniqueSenders: { type: integer, format: uint64 }
        uniqueReceivers: { type: integer, format: uint64 }
        totalActive: { type: integer, format: uint64 }
    NewVsReturningDto:
      type: object
      properties:
        epoch: { type: integer }
        newAddresses: { type: integer, format: uint64 }
        returningAddresses: { type: integer, format: uint64 }
        totalAddresses: { type: integer, format: uint64 }
    ExchangeFlowDto:
      type: object
      properties:
        dataPoints:
          type: array
          items: { $ref: '#/components/schemas/ExchangeFlowDataPointDto' }
        totalInflow: { type: integer, format: uint64 }
        totalOutflow: { type: integer, format: uint64 }
...
    ExchangeSendersDto:
      type: object
      properties:
        clusters:
          type: array
          items: { $ref: '#/components/schemas/AddressClusterDto' }
        senders:
          type: array
          items: { $ref: '#/components/schemas/ExchangeSenderDto' }
        epochsQueried: { type: integer }
        minAmount: { type: number }
...
    HolderDistributionExtendedDto:
      type: object
      properties:
        current: { $ref: '#/components/schemas/HolderDistributionDto' }
        history:
          type: array
          items: { $ref: '#/components/schemas/HolderDistributionHistoryDto' }
...
    NetworkStatsExtendedDto:
      type: object
      properties:
        current:
          nullable: true
          $ref: '#/components/schemas/NetworkStatsHistoryDto'
        history:
          type: array
          items: { $ref: '#/components/schemas/NetworkStatsHistoryDto' }
...